Minor Changes
FillAlpha now works for single image too

+++ FillAlpha.py
@@ -4,9 +4,14 @@ import os
 
 currentPath = str(input('Path to Assets: '))
 
-for FILE in os.listdir(currentPath):
-	if '.py' in FILE or not os.path.isfile(os.path.join(currentPath, FILE)):
-		continue
+if os.path.isfile(currentPath):
+	rgb = cv2.imread(currentPath)
+	cv2.imwrite(currentPath, rgb)
 
-	rgb = cv2.imread(currentPath + '/' + FILE)
-	cv2.imwrite(currentPath + '/' + FILE, rgb)
+else:
+	for FILE in os.listdir(currentPath):
+		if '.py' in FILE or not os.path.isfile(os.path.join(currentPath, FILE)):
+			continue
+
+		rgb = cv2.imread(currentPath + '/' + FILE)
+		cv2.imwrite(currentPath + '/' + FILE, rgb)
